fix: entity info member check is not isolated
The member check did not fall back to check an entity for field-metadata but instead checks all field-metadata from all entities, breaking isolated configuration per entity. It also did not consider array element types.

The `KSqlTypeTranslator` is parameterized by the entity type, this type parameter is used to retrieve entity metadata, to find any field type related configuration for the entity which the translators `Translate` method will translate.

This allows to configure one entity which has a property of some type A with field configuration and also to configure an entity of type A with another set of field configurations. The configurations will not interfere with each other.